stm32_rust_template::driver

Module flash

Structs

Event

Represents events that can occur during flash operations.

FlashInfo

Provides information about the flash memory device. It is assumed that this information is static for a given device.

SectorInfo

Describes a single sector of the flash memory.

Status

Represents the status of the flash memory device.

Traits

Flash

A trait that defines a standard interface for a flash memory driver.

Type Aliases

Error

A generic error type for the flash driver, using i32 for error codes.

Result

A specialized Result type for flash operations.