stm32_rust_template::driver::usart

Trait Usart

pub trait Usart<'a> {

    // Required methods
    fn initialize(
        &mut self,
        callback: impl FnMut(Event) + 'a,
    ) -> Result<(), i32>;
    fn uninitialize(&mut self) -> Result<(), i32>;
    fn configure(&mut self, config: &Config) -> Result<(), i32>;
    fn send(&mut self, data: &[u8]) -> Result<(), i32>;
    fn receive(&mut self, data: &mut [u8]) -> Result<(), i32>;
    fn transfer(
        &mut self,
        data_out: &[u8],
        data_in: &mut [u8],
    ) -> Result<(), i32>;
    fn get_tx_count(&self) -> u32;
    fn get_rx_count(&self) -> u32;
    fn get_status(&self) -> Status;
    fn set_modem_control(&mut self, control: ModemControl) -> Result<(), i32>;
    fn get_modem_status(&self) -> ModemStatus;
    fn tx_enable(&mut self, enable: bool) -> Result<(), i32>;
    fn rx_enable(&mut self, enable: bool) -> Result<(), i32>;
    fn control_break(&mut self, enable: bool) -> Result<(), i32>;
    fn abort_send(&mut self) -> Result<(), i32>;
    fn abort_receive(&mut self) -> Result<(), i32>;
    fn abort_transfer(&mut self) -> Result<(), i32>;

}

A trait that defines a standard interface for a USART driver.

Required Methods

fn initialize(&mut self, callback: impl FnMut(Event) + 'a) -> Result<(), i32>

Initializes the USART peripheral.

The provided callback will be invoked to signal communication events.

fn uninitialize(&mut self) -> Result<(), i32>

De-initializes the USART peripheral.

fn configure(&mut self, config: &Config) -> Result<(), i32>

Configures the USART peripheral.

fn send(&mut self, data: &[u8]) -> Result<(), i32>

Transmits data over the USART bus.

fn receive(&mut self, data: &mut [u8]) -> Result<(), i32>

Receives data from the USART bus.

fn transfer(&mut self, data_out: &[u8], data_in: &mut [u8]) -> Result<(), i32>

Simultaneously sends and receives data.

fn get_tx_count(&self) -> u32

Gets the number of bytes transmitted.

fn get_rx_count(&self) -> u32

Gets the number of bytes received.

fn get_status(&self) -> Status

Gets the current status of the USART peripheral.

fn set_modem_control(&mut self, control: ModemControl) -> Result<(), i32>

Controls the modem lines.

fn get_modem_status(&self) -> ModemStatus

Gets the status of the modem lines.

fn tx_enable(&mut self, enable: bool) -> Result<(), i32>

Enables or disables the transmitter.

fn rx_enable(&mut self, enable: bool) -> Result<(), i32>

Enables or disables the receiver.

fn control_break(&mut self, enable: bool) -> Result<(), i32>

Controls the generation of a break condition.

fn abort_send(&mut self) -> Result<(), i32>

Aborts an ongoing send operation.

fn abort_receive(&mut self) -> Result<(), i32>

Aborts an ongoing receive operation.

fn abort_transfer(&mut self) -> Result<(), i32>

Aborts an ongoing transfer operation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.