stm32_rust_template::mcu

Module stm32f407

STM32F407 MCU Implementation

Contains register definitions, memory maps, and peripheral configurations specific to the STM32F407 microcontroller family.

This module provides base addresses for all peripherals, interrupt numbers, and type-safe register access through the PeripheralAccess trait.

Modules

adc

gpio

i2c

rcc

spi

timer

usart

Enums

IRQn

Constants

ADC1_BASEADDR

ADC2_BASEADDR

ADC3_BASEADDR

ADC_COMMON_BASEADDR

CAN1_BASEADDR

CAN2_BASEADDR

CRC_BASEADDR

CRYP_BASEADDR

DAC_BASEADDR

DBGMCU_BASEADDR

DCMI_BASEADDR

DMA1_BASEADDR

DMA2D_BASEADDR

DMA2_BASEADDR

ETH_DMA_BASEADDR

ETH_MAC_BASEADDR

ETH_MMC_BASEADDR

ETH_PTP_BASEADDR

EXTI_BASEADDR

FLASH_R_BASEADDR

FMC_BASEADDR

FSMC_BASEADDR

GPIOA_BASEADDR

GPIOB_BASEADDR

GPIOC_BASEADDR

GPIOD_BASEADDR

GPIOE_BASEADDR

GPIOF_BASEADDR

GPIOG_BASEADDR

GPIOH_BASEADDR

GPIOI_BASEADDR

GPIOJ_BASEADDR

GPIOK_BASEADDR

HASH_BASEADDR

I2C1_BASEADDR

I2C2_BASEADDR

I2C3_BASEADDR

I2S2EXT_BASEADDR

I2S3EXT_BASEADDR

IWDG_BASEADDR

LTDC_BASEADDR

PWR_BASEADDR

RCC_BASEADDR

RNG_BASEADDR

RTC_BASEADDR

SAI1_BASEADDR

SDIO_BASEADDR

SPI1_BASEADDR

SPI2_BASEADDR

SPI3_BASEADDR

SPI4_BASEADDR

SPI5_BASEADDR

SPI6_BASEADDR

SYSCFG_BASEADDR

TIM1_BASEADDR

TIM2_BASEADDR

TIM3_BASEADDR

TIM4_BASEADDR

TIM5_BASEADDR

TIM6_BASEADDR

TIM7_BASEADDR

TIM8_BASEADDR

TIM9_BASEADDR

TIM10_BASEADDR

TIM11_BASEADDR

TIM12_BASEADDR

TIM13_BASEADDR

TIM14_BASEADDR

UART4_BASEADDR

UART5_BASEADDR

UART7_BASEADDR

UART8_BASEADDR

USART1_BASEADDR

USART2_BASEADDR

USART3_BASEADDR

USART6_BASEADDR

USB_OTG_FS_BASEADDR

USB_OTG_HS_BASEADDR

WWDG_BASEADDR

Traits

PeripheralAccess